requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy<2.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
email-validator==1.1.3
python-dotenv==0.19.0
Flask-Mail==0.9.1
Flask-Limiter==1.4
email_validator==1.1.3
werkzeug==2.2.2
wtforms==2.3.3
tensorflow-cpu
textblob
scikit-learn
numpy
pillow